msm: mdss: hdmi: Avoid unnecessary BCAPS read
Don't read BCAPS (downstream device's capability register) in second
part of authentication to check if the attached downstream device is
a repeater or not. BCAPS register is already read in first part of
authentication. Reuse it's value for this purpose. This avoids any
unnecessary delays and possible arbitration for DDC line between
hardware (for third part of authentication) and software which
sometimes results in HDCP compliance 1A-07 failure or not judged.

Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>