Revert "crypto: Fix memory leak in bam_release call"
This reverts commit 86b9fe8005f9bd6d245f1e3c6596d3e4549b0e6f, which
release pbam memory when an error occurs. However, on a device with
multiple ce, pbam memory is shared by multiple pce_dev, it can only
be freed when the reference count goes to zero, and can not be
released it if reference count is not zero when an error occurs.
The previous implementation on bam_release was correct, so we
revert commit 86b9fe8005f9bd6d245f1e3c6596d3e4549b0e6f.

Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>